@@ -1,16 +1,15 @@
-﻿namespace FastHashesNet.FarmHash
+﻿namespace FastHashesNet.FarmHash;
+
+public struct Uint128
 {
-    public struct Uint128
+    public Uint128(ulong low, ulong high)
+        : this()
     {
-        public Uint128(ulong low, ulong high)
-            : this()
-        {
-            Low = low;
-            High = high;
-        }
-
-        public ulong Low { get; set; }
-
-        public ulong High { get; set; }
+        Low = low;
+        High = high;
     }
+
+    public ulong Low { get; set; }
+
+    public ulong High { get; set; }
 }
