@@ -1,11 +1,10 @@
-﻿namespace FastHashesNet.MurmurHash
-{
-    internal static class MurmurHashConstants
-    {
-        internal const uint C1_32 = 0xcc9e2d51;
-        internal const uint C2_32 = 0x1b873593;
+﻿namespace FastHashesNet.MurmurHash;
 
-        internal const ulong C1_64 = 0x87c37b91114253d5UL;
-        internal const ulong C2_64 = 0x4cf5ad432745937fUL;
-    }
+internal static class MurmurHashConstants
+{
+    internal const uint C1_32 = 0xcc9e2d51;
+    internal const uint C2_32 = 0x1b873593;
+
+    internal const ulong C1_64 = 0x87c37b91114253d5UL;
+    internal const ulong C2_64 = 0x4cf5ad432745937fUL;
 }
