@@ -1,10 +1,9 @@
-﻿namespace FastHashesNet.SipHash
+﻿namespace FastHashesNet.SipHash;
+
+public static class SipHashConstants
 {
-    public static class SipHashConstants
-    {
-        public const ulong v0Init = 0x736f6d6570736575U;
-        public const ulong v1Init = 0x646f72616e646f6dU;
-        public const ulong v2Init = 0x6c7967656e657261U;
-        public const ulong v3Init = 0x7465646279746573U;
-    }
+    public const ulong v0Init = 0x736f6d6570736575U;
+    public const ulong v1Init = 0x646f72616e646f6dU;
+    public const ulong v2Init = 0x6c7967656e657261U;
+    public const ulong v3Init = 0x7465646279746573U;
 }
